+++ com.unity.cinemachine/Runtime/Core/SplineHelpers.cs
@@ -1,3 +1,4 @@
+using Unity.Mathematics;
 using UnityEngine;
 
 namespace Unity.Cinemachine
@@ -182,6 +183,116 @@ namespace Unity.Cinemachine
             Vector4[] knotLooped = new Vector4[numPoints + 2 * margin];
             Vector4[] ctrl1Looped = new Vector4[numPoints + 2 * margin];
             Vector4[] ctrl2Looped = new Vector4[numPoints + 2 * margin];
+            for (int i = 0; i < margin; ++i)
+            {
+                knotLooped[i] = knot[numPoints-(margin-i)];
+                knotLooped[numPoints+margin+i] = knot[i];
+            }
+            for (int i = 0; i < numPoints; ++i)
+                knotLooped[i + margin] = knot[i];
+            ComputeSmoothControlPoints(ref knotLooped, ref ctrl1Looped, ref ctrl2Looped);
+            for (int i = 0; i < numPoints; ++i)
+            {
+                ctrl1[i] = ctrl1Looped[i + margin];
+                ctrl2[i] = ctrl2Looped[i + margin];
+            }
+        }
+
+        /// <summary>
+        /// Use the Thomas algorithm to compute smooth tangent values for a spline.  
+        /// Resultant tangents guarantee second-order smoothness of the curve.
+        /// </summary>
+        /// <param name="knot">The knots defining the curve</param>
+        /// <param name="ctrl1">Output buffer for the first control points (1 per knot)</param>
+        /// <param name="ctrl2">Output buffer for the second control points (1 per knot)</param>
+        public static void ComputeSmoothControlPoints(
+            ref float3[] knot, ref float3[] ctrl1, ref float3[] ctrl2)
+        {
+            int numPoints = knot.Length;
+            if (numPoints <= 2)
+            {
+                if (numPoints == 2)
+                {
+                    ctrl1[0] = math.lerp(knot[0], knot[1], 0.33333f);
+                    ctrl2[0] = math.lerp(knot[0], knot[1], 0.66666f);
+                }
+                else if (numPoints == 1)
+                    ctrl1[0] = ctrl2[0] = knot[0];
+                return;
+            }
+
+            var a = new float[numPoints];
+            var b = new float[numPoints];
+            var c = new float[numPoints];
+            var r = new float[numPoints];
+            for (int axis = 0; axis < 3; ++axis)
+            {
+                int n = numPoints - 1;
+
+                // Linear into the first segment
+                a[0] = 0;
+                b[0] = 2;
+                c[0] = 1;
+                r[0] = knot[0][axis] + 2 * knot[1][axis];
+
+                // Internal segments
+                for (int i = 1; i < n - 1; ++i)
+                {
+                    a[i] = 1;
+                    b[i] = 4;
+                    c[i] = 1;
+                    r[i] = 4 * knot[i][axis] + 2 * knot[i+1][axis];
+                }
+
+                // Linear out of the last segment
+                a[n - 1] = 2;
+                b[n - 1] = 7;
+                c[n - 1] = 0;
+                r[n - 1] = 8 * knot[n - 1][axis] + knot[n][axis];
+
+                // Solve with Thomas algorithm
+                for (int i = 1; i < n; ++i)
+                {
+                    float m = a[i] / b[i-1];
+                    b[i] = b[i] - m * c[i-1];
+                    r[i] = r[i] - m * r[i-1];
+                }
+
+                // Compute ctrl1
+                ctrl1[n-1][axis] = r[n-1] / b[n-1];
+                for (int i = n - 2; i >= 0; --i)
+                    ctrl1[i][axis] = (r[i] - c[i] * ctrl1[i + 1][axis]) / b[i];
+
+                // Compute ctrl2 from ctrl1
+                for (int i = 0; i < n; i++)
+                    ctrl2[i][axis] = 2 * knot[i + 1][axis] - ctrl1[i + 1][axis];
+                ctrl2[n - 1][axis] = 0.5f * (knot[n][axis] + ctrl1[n - 1][axis]);
+            }
+        }
+
+        /// <summary>
+        /// Use the Thomas algorithm to compute smooth tangent values for a spline.  
+        /// This method will assume that the spline is looped (i.e. that the last knot is followed by the first).
+        /// Resultant tangents guarantee second-order smoothness of the curve.
+        /// </summary>
+        /// <param name="knot">The knots defining the curve</param>
+        /// <param name="ctrl1">Output buffer for the first control points (1 per knot)</param>
+        /// <param name="ctrl2">Output buffer for the second control points (1 per knot)</param>
+        public static void ComputeSmoothControlPointsLooped(
+            ref float3[] knot, ref float3[] ctrl1, ref float3[] ctrl2)
+        {
+            int numPoints = knot.Length;
+            if (numPoints < 2)
+            {
+                if (numPoints == 1)
+                    ctrl1[0] = ctrl2[0] = knot[0];
+                return;
+            }
+
+            int margin = Mathf.Min(4, numPoints-1);
+            float3[] knotLooped = new float3[numPoints + 2 * margin];
+            float3[] ctrl1Looped = new float3[numPoints + 2 * margin];
+            float3[] ctrl2Looped = new float3[numPoints + 2 * margin];
             for (int i = 0; i < margin; ++i)
             {
                 knotLooped[i] = knot[numPoints-(margin-i)];